Moved clearDownloads to actions/Download.js
