+++ src/libs/actions/Download.js
@@ -1,3 +1,4 @@
+import _ from 'underscore';
 import Onyx from 'react-native-onyx';
 import ONYXKEYS from '../../ONYXKEYS';
 
@@ -12,7 +13,22 @@ function setDownload(sourceID, isDownloading) {
     return Onyx.merge(`${ONYXKEYS.COLLECTION.DOWNLOAD}${sourceID}`, {isDownloading});
 }
 
+function clearDownloads() {
+    const connectionID = Onyx.connect({
+        key: ONYXKEYS.COLLECTION.DOWNLOAD,
+        waitForCollectionCallback: true,
+        callback: (records) => {
+            Onyx.disconnect(connectionID);
+            const downloads = {};
+            _.each(_.keys(records), recordKey => downloads[recordKey] = null);
+            if (!_.isEmpty(downloads)) {
+                Onyx.multiSet(downloads);
+            }
+        },
+    });
+}
+
 export {
-    // eslint-disable-next-line import/prefer-default-export
     setDownload,
+    clearDownloads,
 };
